{voduges} fixed warnings, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/branches/sharps@19597 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ src/microsim/MSEdge.cpp
@@ -124,7 +124,7 @@ MSEdge::initialize(const std::vector<MSLane*>* lanes) {
     if (MSGlobals::gLateralResolution > 0 || MSGlobals::gLaneChangeDuration > 0) {
         SUMOReal widthBefore = 0;
         for (std::vector<MSLane*>::const_iterator i = myLanes->begin(); i != myLanes->end(); ++i) {
-            (*i)->setRightSideOnEdge(widthBefore, mySublaneSides.size());
+            (*i)->setRightSideOnEdge(widthBefore, (int)mySublaneSides.size());
             MSLeaderInfo ahead(*i);
             for (int j = 0; j < ahead.numSublanes(); ++j) {
                 mySublaneSides.push_back(widthBefore + j * MSGlobals::gLateralResolution);
